requirements.txt
numpy==1.24.3
pandas==2.0.1
python-dateutil==2.8.2
pytube==15.0.0
pytz==2023.3
six==1.16.0
tzdata==2023.3
moviepy==1.0.3
praw==7.7.0
gTTS==2.2.3
python-dotenv==1.0.0